--- src/v2/ObjectTransformer/ValueHandlerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Itk\EventDatabaseClient\v2\ObjectTransformer;
-
-interface ValueHandlerInterface
-{
-    public function getValue($item, $path);
-
-    public function convertValue($value, $key);
-}
